Unittests can now be 100% automatic other than launching RethinkDB
Added root access key generation script
Unittests now automatically get the root token and use it for testing

+++ tests/conftest.py
@@ -1,33 +1,53 @@
 import sys
 import pytest
+import subprocess
 import rethinkdb as rethink
 from os import environ as env
-import subprocess
 
 # Why this works, I don't know. Blech.
 sys.path.append(".")
+sys.path.append("..")
+
+from app.util import helpers
+
+global token
 
 
 def pytest_sessionstart(session):
-    # Generate root access token (gen.py --raw -p sys.argv[2] -u sys.argv[1])
+    # Generate root access token
+    from app import app
+    if "ROOT_PASSWORD" not in app.config:
+        password = input("Input the root password: ")
+    else:
+        password = app.config["ROOT_PASSWORD"]
 
-    if "APIKEY" not in env:
-        pytest.exit("NEED THE API ACCESS KEY YA FOOL!")
+    root_pass = subprocess.run("python3 gen.py --raw {password}".format(
+        password=password
+    ), stdout=subprocess.PIPE, shell=True)
+
+    if root_pass.stdout.decode().rstrip() == "password":
+        pytest.exit("\033[91mPASSWORD DOES NOT MATCH ROOT PASSWORD\033[0m")
+    elif root_pass.stdout.decode().rstrip() == "account":
+        pytest.exit("\033[91mROOT ACCOUNT IS NOT CONFIGURED\033[0m")
+    else:
+        global token
+        token = root_pass.stdout.decode().rstrip()
 
 
 def pytest_sessionfinish(session):
     r = rethink.connect(db="testing")
     for key in rethink.db("testing").table_list().run(r):
-        rethink.db("testing").table(key).delete().run(r)
+        if key not in ("keys", "users"):
+            rethink.db("testing").table(key).delete().run(r)
 
 
 @pytest.fixture
 def app(request):
-    from app import app
+    global token
     app.config["TESTING"] = True
     app.config["RDB_DB"] = "testing"
     app.config["API_TOKEN"] = "root"
-    app.config["API_KEY"] = env["APIKEY"]
+    app.config["API_KEY"] = token
 
     return app
